+++ src/main/java/micycle/pgs/PGS_Processing.java
@@ -54,8 +54,8 @@ import micycle.balaban.BalabanSolver;
 import micycle.balaban.Point;
 import micycle.balaban.Segment;
 import micycle.pgs.PGS.GeometryIterator;
-import micycle.pgs.utility.PolygonDecomposition;
-import micycle.pgs.utility.SeededRandomPointsInGridBuilder;
+import micycle.pgs.commons.PolygonDecomposition;
+import micycle.pgs.commons.SeededRandomPointsInGridBuilder;
 import processing.core.PConstants;
 import processing.core.PShape;
 import processing.core.PVector;
