+++ src/main/java/micycle/pgs/commons/DoyleSpiral.java
@@ -1,4 +1,4 @@
-package micycle.pgs.utility;
+package micycle.pgs.commons;
 
 import static java.lang.Math.pow;
 import static java.lang.Math.sin;
