+++ src/main/java/micycle/pgs/commons/FrontChainPacker.java
@@ -1,4 +1,4 @@
-package micycle.pgs.utility;
+package micycle.pgs.commons;
 
 import java.util.ArrayList;
 import java.util.List;
