+++ src/main/java/micycle/pgs/commons/MaximumInscribedRectangle.java
@@ -1,4 +1,4 @@
-package micycle.pgs.utility;
+package micycle.pgs.commons;
 
 import java.awt.Point;
