+++ src/main/java/micycle/pgs/commons/MinimumBoundingTriangle.java
@@ -1,4 +1,4 @@
-package micycle.pgs.utility;
+package micycle.pgs.commons;
 
 import static java.lang.Math.floorMod;
