+++ src/main/java/micycle/pgs/commons/Nullable.java
@@ -1,4 +1,4 @@
-package micycle.pgs.utility;
+package micycle.pgs.commons;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
