+++ src/main/java/micycle/pgs/commons/PoissonDistribution.java
@@ -1,4 +1,4 @@
-package micycle.pgs.utility;
+package micycle.pgs.commons;
 
 import static processing.core.PApplet.ceil;
 import static processing.core.PApplet.floor;
