+++ src/main/java/micycle/pgs/commons/PolygonDecomposition.java
@@ -22,7 +22,7 @@
  * IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT 
  * OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package micycle.pgs.utility;
+package micycle.pgs.commons;
 
 import java.util.ArrayList;
 import java.util.Collections;
