+++ src/main/java/micycle/pgs/commons/RLFColoring.java
@@ -1,4 +1,4 @@
-package micycle.pgs.utility;
+package micycle.pgs.commons;
 
 import java.util.HashSet;
 import java.util.Map;
