+++ src/main/java/micycle/pgs/commons/package-info.java
@@ -3,4 +3,4 @@
  * and either not worth putting in separate package or too long to include
  * within the calling PGS method itself).
  */
-package micycle.pgs.utility;
+package micycle.pgs.commons;
